Do more sample things
