+++ src/Controls/samples/Controls.Sample.Sandbox/App.xaml.cs
@@ -1,0 +1,24 @@
+﻿namespace Maui.Controls.Sample;
+
+public partial class App : Application
+{
+	public App()
+	{
+		InitializeComponent();
+	}
+
+	protected override Window CreateWindow(IActivationState? activationState)
+	{
+		// To test shell scenarios, change this to true
+		bool useShell = false;
+
+		if (!useShell)
+		{
+			return new Window(new NavigationPage(new MainPage()));
+		}
+		else
+		{
+			return new Window(new SandboxShell());
+		}
+	}
+}
