+++ src/Controls/samples/Controls.Sample.Sandbox/MainPage.xaml
@@ -3,4 +3,5 @@
     xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
     x:Class="Maui.Controls.Sample.MainPage"
     xmlns:local="clr-namespace:Maui.Controls.Sample">
+
 </ContentPage>
