+++ src/Controls/samples/Controls.Sample.Sandbox/MainPage.xaml.cs
@@ -1,18 +1,9 @@
-﻿using System;
-using System.Collections.ObjectModel;
-using System.Diagnostics;
-using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Maui;
-using Microsoft.Maui.Controls;
-using Microsoft.Maui.Graphics;
+﻿namespace Maui.Controls.Sample;
 
-namespace Maui.Controls.Sample
+public partial class MainPage : ContentPage
 {
-	public partial class MainPage : ContentPage
+	public MainPage()
 	{
-		public MainPage()
-		{
-			InitializeComponent();
-		}
+		InitializeComponent();
 	}
 }
