+++ src/Controls/samples/Controls.Sample.Sandbox/SandboxShell.xaml.cs
@@ -1,17 +1,9 @@
-﻿using System;
-using System.Collections.ObjectModel;
-using System.Diagnostics;
-using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Maui;
-using Microsoft.Maui.Controls;
+﻿namespace Maui.Controls.Sample;
 
-namespace Maui.Controls.Sample
+public partial class SandboxShell : Shell
 {
-	public partial class SandboxShell : Shell
+	public SandboxShell()
 	{
-		public SandboxShell()
-		{
-			InitializeComponent();
-		}
+		InitializeComponent();
 	}
 }
